requirements.txt
google-generativeai   # libreria de google para IA
